refactor(frontend): shared data params state
Move generic data params state into `src/lib`. The aim here is that the app should be responsible for loading data params config, but the shared `lib` package should be responsible for handling data params state and updating the map.
