+++ frontend/src/app/sidebar/ui/params/DataParam.tsx
@@ -1,5 +1,5 @@
 import { useRecoilValue } from 'recoil';
-import { dataParamOptionsState, dataParamState, useUpdateDataParam } from 'app/state/data-params';
+import { dataParamOptionsState, dataParamState, useUpdateDataParam } from 'lib/state/data-params';
 
 export const DataParam = ({ group, id, children }) => {
   const value = useRecoilValue(dataParamState({ group, param: id }));
