+++ frontend/src/data-layers/networks/state/layer.ts
@@ -4,11 +4,12 @@ import fromPairs from 'lodash/fromPairs';
 import mapValues from 'lodash/mapValues';
 
 import { recalculateCheckboxStates } from 'lib/controls/checkbox-tree/CheckboxTree';
-import { LayerSpec } from 'app/asset-list/use-sorted-features';
 import { ViewLayer, StyleParams, ColorSpec, FieldSpec } from 'lib/data-map/view-layers';
 import { StateEffect } from 'lib/recoil/state-effects/types';
+import { dataParamsByGroupState } from 'lib/state/data-params';
+
+import { LayerSpec } from 'app/asset-list/use-sorted-features';
 import { damageMapStyleParamsState } from 'app/state/damage-mapping/damage-style-params';
-import { dataParamsByGroupState } from 'app/state/data-params';
 import { sectionVisibilityState } from 'app/state/sections';
 
 import adaptationSectorLayers from '../adaptation-sector-layers.json';
