+++ frontend/src/details/adaptations/AdaptationsSidebar.tsx
@@ -1,16 +1,21 @@
 import { Typography } from '@mui/material';
 import { Box } from '@mui/system';
-import { SidePanel } from 'details/SidePanel';
-import { ErrorBoundary } from 'lib/react/ErrorBoundary';
 import { FC } from 'react';
-import { FeatureAdaptationsTable } from './FeatureAdaptationsTable';
+
+import { ErrorBoundary } from 'lib/react/ErrorBoundary';
 import { MobileTabContentWatcher } from 'lib/map/layouts/tab-has-content';
+
+import { useSyncConfigState } from 'app/state/data-params';
+import { SidePanel } from 'details/SidePanel';
+
+import { FeatureAdaptationsTable } from './FeatureAdaptationsTable';
 
 /**
  * List adaptation options for a selected infrastructure network
  * eg. power lines, roads, railway tracks etc.
  */
 export const AdaptationsSidebar: FC = () => {
+  useSyncConfigState();
   return (
     <SidePanel height="80vh" pb={1} px={0} pt={0}>
       <MobileTabContentWatcher tabId="details" />
